feat(PatternFuncProcessor): change argument addition order and improve error message
This commit changes the order of adding arguments in the PatternFuncProcessor. Instead of adding the lastResult at the beginning of the args list, it is now added at the end. Additionally, the error message for unmatched functions has been improved to provide more information and a link to the toolchain function documentation. A new setting to focus the tool window before running has also been added in the WiremockFunctionProvider.