feat(java-testing): add syntax error collection and auto testing service for Java
This commit introduces a new feature that allows the collection of syntax errors in Kotlin files and the creation of an automatic testing service for Java. It also includes the necessary updates to the plugin version and the addition of the JavaAutoTestingService to the shireAutoTesting language in the XML configuration file.