feat(run-service): add isApplicable method to FileRunService
Added an `isApplicable` method to the `FileRunService` interface and implemented it in various classes. This method checks if a given project and file are applicable for the service. This change improves the flexibility and extensibility of the run service, allowing it to handle a wider range of file types and project configurations.