refactor(middleware): Enable resizing and moving for webview popup
Allows users to resize and move the webview popup for better usability. Removes the commented out code for setting a minimum size, as it was not functional.

The popup created by `ShowWebviewProcessor.kt` now has the capabilities to be resized and moved, enhancing the user experience. This change does not set a fixed minimum size, providing more flexibility to the users.