refactor(project): rename and update openwrite to openrewrite #119
This commit renames the `openwrite` project to `openrewrite` and updates all references accordingly. The changes include updating build and settings files, renaming source files, and adjusting plugin.xml and project configuration files to reflect the new name. Additionally, there are improvements to the `OpenRewriteFileRunService` class, including a more specific check for applicable files and updates to the run configuration process.