refactor: move codemodel classes to provider package
This commit involves the refactoring of core codemodel classes to a new package called 'provider'. The main objective is to reorganize the classes under the 'codemodel' package to a more structured 'provider' package. This includes renaming the classes and updating the imports accordingly. The affected classes include:

- ClassStructureProvider
- FileStructureProvider
- MethodStructureProvider
- VariableStructureProvider
- FileStructure
- ClassStructure
- MethodStructure
- VariableStructure
- DirectoryStructure

This refactoring does not alter the functionality but enhances the organization of the codebase.